fix(ide-debugger): 显式 cast lastVisibleRow 避免 float→int lossy 转换
`editor.getFirstVisibleRow()` 返回 float,所以
`firstVisibleRow + visibleRowCount + 1` 整个表达式是 float,
不能直接赋给 `int lastVisibleRow`。加 `(int)` 显式 cast。

+++ core/app/src/main/java/com/itsaky/androidide/debugger/view/BreakpointSidebar.java
@@ -131,9 +131,11 @@ public class BreakpointSidebar extends View {
         // Sora editor doesn't expose a `getRowCountOnScreen()` helper, so
         // approximate it from the editor's height and the row height. We
         // add a 1-row slack for the partially visible trailing row.
+        // `getFirstVisibleRow()` returns a float, so the final addition is
+        // also a float; cast back to int for the row index comparison.
         final int visibleRowCount = (int) Math.ceil(
                 editor.getHeight() / Math.max(1f, editor.getRowHeight()));
-        final int lastVisibleRow = firstVisibleRow + visibleRowCount + 1;
+        final int lastVisibleRow = (int) (firstVisibleRow + visibleRowCount + 1);
 
         final float cx = getWidth() / 2f;
         final float r = dp(GLYPH_RADIUS_DP);
